+++ packages/sui-ssr/server/utils/factory.js
@@ -6,8 +6,9 @@ const DEFAULT_MULTI_SITE_KEY = 'default'
 const EXPRESS_STATIC_CONFIG = {index: false}
 
 let cachedCriticalManifest
+let cachedAssetsManifest
 
-export default ({path, fs, config: ssrConf = {}}) => {
+export default ({path, fs, config: ssrConf = {}, assetsManifest}) => {
   const multiSiteMapping = ssrConf.multiSite
   const multiSiteKeys = multiSiteMapping && Object.keys(multiSiteMapping)
 
@@ -114,23 +115,52 @@ export default ({path, fs, config: ssrConf = {}}) => {
     return url
   }
 
-  const getStyleHrefBy = ({pageName}) => {
-    const assetsFile = ssrConf.assetsManifest
+  const getAssetsManifest = ({req}) => {
+    const site = siteByHost(req)
+
+    if (cachedAssetsManifest) {
+      return site ? cachedAssetsManifest[site] : cachedAssetsManifest
+    }
+
+    let assetsManifest
+
+    try {
+      assetsManifest = JSON.parse(
+        fs.readFileSync(
+          path.join(process.cwd(), publicFolder(req), 'asset-manifest.json'),
+          'utf8'
+        )
+      )
+      if (site) {
+        cachedAssetsManifest = {
+          ...cachedAssetsManifest,
+          [site]: assetsManifest
+        }
+      } else {
+        cachedAssetsManifest = assetsManifest
+      }
+    } catch (error) {
+      assetsManifest = null
+    }
+
+    return assetsManifest
+  }
+
+  const getStyleHrefBy = ({pageName, req}) => {
+    const assetsFile = assetsManifest || getAssetsManifest({req})
 
     return assetsFile && assetsFile[`${pageName}.css`]
   }
 
-  const createStylesFor = ({pageName, async = false} = {}) => {
+  const createStylesFor = ({pageName, async = false, req} = {}) => {
     if (!ssrConf.createStylesFor) return ''
 
     const appStyles = ssrConf.createStylesFor.appStyles
     const shouldCreatePageStyles = ssrConf.createStylesFor.createPagesStyles
-
     const stylesheets = [
-      appStyles && getStyleHrefBy({pageName: appStyles}),
-      shouldCreatePageStyles && getStyleHrefBy({pageName})
+      appStyles && getStyleHrefBy({pageName: appStyles, req}),
+      shouldCreatePageStyles && getStyleHrefBy({pageName, req})
     ].filter(Boolean)
-
     const attributes = async ? ssrConf.ASYNC_CSS_ATTRS : ''
     const stylesHTML = stylesheets
       .map(style => `<link rel="stylesheet" href="${style}" ${attributes}>`)
@@ -167,6 +197,7 @@ export default ({path, fs, config: ssrConf = {}}) => {
 
   return {
     buildRequestUrl,
+    getAssetsManifest,
     createStylesFor,
     hostFromReq,
     hrTimeToMs,